Registro - 002 - Entidades implementadas en BD (QA)
