+++ feel-flow-package/src/main/java/com/equipo5/feelflowapp/domain/RegularUser.java
@@ -1,0 +1,17 @@
+package com.equipo5.feelflowapp.domain;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.PrimaryKeyJoinColumn;
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@PrimaryKeyJoinColumn(name = "regularuser_persona")
+public class RegularUser extends User{
+    @ManyToOne
+    private Team team;
+}
