+++ feel-flow-package/src/main/java/com/equipo5/feelflowapp/domain/TeamLeader.java
@@ -1,0 +1,18 @@
+package com.equipo5.feelflowapp.domain;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.PrimaryKeyJoinColumn;
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@PrimaryKeyJoinColumn(name = "teamleader_persona")
+public class TeamLeader extends User{
+
+    @OneToOne(mappedBy = "teamLeader")
+    private Team team;
+}
