--- feel-flow-package/src/main/java/com/equipo5/feelflowapp/domain/users/Admin.java
@@ -1,13 +1,0 @@
-package com.equipo5.feelflowapp.domain.users;
-
-import com.equipo5.feelflowapp.domain.EnterPrise;
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Admin extends User{
-    private EnterPrise enterPrise;
-}
