--- feel-flow-package/src/main/java/com/equipo5/feelflowapp/domain/users/TeamLeader.java
@@ -1,13 +1,0 @@
-package com.equipo5.feelflowapp.domain.users;
-
-import com.equipo5.feelflowapp.domain.Team;
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class TeamLeader extends User{
-    private Team team;
-}
